requirements.txt
django
djangorestframework
psycopg2
flake8
flake8-bugbear
pep8-naming